core: do not add child views in the layer list
We should really treat them as child views. Their position is fully
detemined by the Z order of the main view.

Additional changes and motivation for them:

1. attach/detach/layer-attach/layer-detach signals needed to be
clarified, because now we can detach views which are no longer in any
layer.

2. Split view initialization in 3 stages. This is needed because
otherwise we'd be attaching the view to an output before it is allocated
at all.

3. Focusing a view now has a somehow different semantics. When a view
is focused, this means that its frontmost child (if no children, the
view itself) gets focus. As a result, we need to refocus the view each
time the parent-child relation changes.